move donor ref analysis to separate file
